requirements.txt
backoff==1.8.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
ciso8601==2.1.3
click==7.1.1
cssselect==1.1.0
idna==2.9
jsonschema==2.6.0
lxml==4.5.0
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2018.4
requests==2.23.0
simplejson==3.11.1
singer-python==5.9.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
